[ALPS06060264] cam: mt6983: IMX766 fix dual exp gain
Fix data type of gain.

MTK-Commit-Id: d3f8780bc4166b3faa022f5d8eef3256c54b6e25

CR-Id: ALPS06060264
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/imgsensor/src-v4l2/common/imx766dual_mipi_raw/imx766dualmipiraw_Sensor.c
@@ -3464,12 +3464,12 @@ static int feature_control(struct subdrv_ctx *ctx, MSDK_SENSOR_FEATURE_ENUM feat
 		break;
 	case SENSOR_FEATURE_SET_DUAL_GAIN://for 2EXP
 		LOG_INF("SENSOR_FEATURE_SET_DUAL_GAIN LE=%d, SE=%d\n",
-				(UINT16) *feature_data, (UINT16) *(feature_data + 1));
+				(UINT32)*feature_data, (UINT32)*(feature_data + 1));
 		// implement write gain for NE/SE
 		hdr_write_tri_gain(ctx,
-				(UINT32) * feature_data,
+				(UINT32)*feature_data,
 				0,
-				(UINT32) * (feature_data+1));
+				(UINT32)*(feature_data+1));
 		break;
 	case SENSOR_FEATURE_SET_HDR_TRI_SHUTTER://for 3EXP
 		LOG_INF("SENSOR_FEATURE_SET_HDR_TRI_SHUTTER LE=%d, ME=%d, SE=%d\n",
